Remove boolean from ShorthandProps type (#18521)
Remove boolean as an allowed type of ShorthandProps.

#### Background
Currently setting a shorthand slot to `true` or `false` will render the slot's element, but not render any children. This is confusing and possibly not what one would expect. It would be reasonable to expect that setting a slot to `false` would cause it to not render at all (which isn't the case). And there's no reasonable expectation of what would happen when setting the slot to true.

For example, `<Button icon={false}>...</Button>` will render the icon slot with nothing in it, and the button will have a blank space for the icon. This is likely not what was intended.

Also see the change in Avatar.stories.tsx in this PR to see an example of a boolean value in a slot causing incorrect behavior.

Given the confusing nature of having a shorthand be boolean, and questionable benefits, it seems best to remove boolean as a possible type for ShorthandProps. If there is some legitimate reason to have the slot's children be a boolean, it can still be written "longhand" as `slot={{ children: false }}`