Fix crash in ContextualMenu when Callout component is loaded asynchronously (#19122)
PR #15625 moved the `hostElement` ref from an inner `div` of ContextualMenu, to the root `Callout` element. If the Callout component is loaded asynchronously, this can cause a crash when the ref doesn't refer to a fully loaded element.

This change un-merges the `forwardedRef` and `hostElement` ref. It still applies the forwarded ref to the root Callout (to maintain the fix from #15625), but puts the `hostElement` ref  back to the inner `div`.