Add IntrinsicShorthandProps type for slots that use intrinsic elements like div or span (#19642)
Introduce a new type IntrinsicShorthandProps for slots that use intrinsic elements like div or span. This replaces "As" on ObjectShorthandProps, and gives a more ergonomic way of specifying slots props that support multiple element types. This is intended to be the main way to specify slot props for slots that are an intrinsic element type.