Rename Radio to RadioGroup, and RadioItem to Radio (#21817)
Rename the react-radio components to `RadioGroup` for the group, and `Radio` for the individual item. This aligns the names with the ARIA roles `radiogroup` and `radio`, as well as the input type `radio`.

Note: The diff in this change is misleading. The only changes were renames of Radio => RadioGroup, and RadioItem => Radio, plus updates to the spec. Because the old name of RadioGroup (Radio) is the same as the new name for RadioItem (Radio), they appear as edits rather than renames.