SpinButton implementation and styles (#22233)
* First pass on SpinButton implementation and styles

* Fix lint and tests.

* Update types docs to fix linting warning

* Update isConformant for react-spinbutton

* Export missing SpinButton types

* Change @default to @defaultValue

* react-spinbutton: update styling implementation

Rework the class name for active buttons to be simpler.

* react-spinbutton: add extra case for clamp function

clampWhenInRange() now handles the case when the min value is greater
than the max value. In this case the new value is simply returned with
no clamping applied.

In dev mode an error is written to the console as this is an invalid
state.

Tests have been added to verify the behavior.

* react-spinbutton: add Enter key support

Adds support for the Enter key to SpinButton. When focused on the text
input field hitting the enter key will now commit the value.

* react-spinbutton: remove size prop

Size prop was being passed down to the input slot. This was a left over
from attempting to use `Input` rather than `input` for this slot. It is
no longer needed.