react-input and react-spinbutton use ::after pseudo element (#22591)
Updates the styles in Input and SpinButton to use the `::after`
pseudo element rather than `:after` as `::after` is the preferred
notation in CSS3 while `:after` is the CSS2 syntax. Both are valid and
acceptable but we should standardize on the newer syntax as we now only
support evergreen browsers.