list: add utilities for reading and writing scroll values (#23813)
* list: add utilities for reading and writing scroll values

This commit adds three scrolling utilities:

1. getScrollHeight: to read the scroll height of a List parent
2. getScrollYPosition: to read the scroll position of a List parent
3. setScrollYPosition: to set the Y scroll position of a List parent

The scroll parent of a List can be either an HTMLElement or the Window
object which is a problem because these two objects have different APIs
for working with scroll position. The new scrolling utilities handle
these API differences internally.

The prior implementation assumed the HTMLElement API and didn't support
the Window API. This has worked out in practice because the existing
implementation does not throw errors but it does mean that there are
cases where values like `0` or `undefined` are always returned when
working with scroll positioning. These issues only come up when the
List's scroll parent is Window which is probably an unlikely situation
in an actual application but it definitely arises during development
and debugging when List is used in isolation.

* list: reorganize utils tests