feat: Remove inferred row state and add `rowEnhancer` option (#24346)
* feat: Remove inferred row state and add `rowEnhancer` option

A lot of the row state was actually inferred from other state. This
creates an unnecessary coupling between the selection state and the
actual render data.

To impreove this, this PR adds the `rowEnhancer` concept. This allows
users to 'enhance' their rows with extra functionality or data with
access to the table state. The enhancer runs during the creation of
`rows` which means the user does not need to unnecessarily iterate over
the return of `useTable` again.

Also makes it easier to split up the different substates in the future

* update API

* changefile

* simplify sort state

* row enhancer with public state

* update api.md

* update tests