[v8] fix: Make focus rects visible again in Combobox (#24340)
FocusRectProvider now keeps an array of all child provider elements in the React tree, including ones in a separate Layer. When focus visibility changes, the styles are applied on the current and all child providers.

The net effect is that keyboard events on the root will affect any open Layers/Callouts.