stress test: convert cli scripts to typescript (#24915)
This change gives us better type checking (scripts are now covered by
the `type-check` command) while also allowing interop with the source
code. For example, we can add a command to generate fixtures from the
`RandomTree` class. This was not possible when the scripts were written
in JS as the source code is otherwise in Typescript.