stress-test: add ability to target any url (#25281)
Prior to this change only pages hosted by the Stress Test server could
be tested. This change allows for targets to be pointed at other servers
by using a target name that starts with `http` (e.g., http://example.com/page-to-test).

This is particularly useful when testing aginst divergent forks of Fluent
where you want to test against the current state of the world and a
proposed new state.