fix: list even/odd off-by-one issue (#25358)
Fixes an issue in List's scrolling example where even and odd were
calculated differently in CSS and Javascript. JS was using a 0 index
while CSS was using a 1 index so height calculations were off by the
height of an element in the List. This was only apparent in lists were
items had different heights as in the example.