GroupedListV2: update aria-posinset and aria-rowindex (#25520)
* groupedlistv2: update aria positioning attributes

Changes the way aria-posinset is computed when the GroupedList role is
set to "treegrid", the default role for GroupedList. In this case
aria-posinset is the index within the current set.

In all other cases only aria-rowindex is set. This case specifically is
used in grouped DetailsList.

* groupedlistv2: add detailslist examples

Add two DetailsList examples that use GroupedListV2 under the hood.
Update the doc site to show these examples.

* dl

* groupedlistv2: properly set aria values for non-expandable rows

GroupedListV1 case `onRenderCell` to `any`, allowing an additional
`group` prop to be passed to this render callback. This updates
`IGroupedListV2Props` to include this as an optional argument to the
`onRenderCell` callback.

Passing `group` is necessary as it holds the information needed to set
`aria-posinset` and `aria-setsize` for non-expandable rows.

* groupedlistv2: make some props optional

* Apply suggestions from code review

Co-authored-by: Esteban Munoz Facusse <esteban.230@hotmail.com>

Co-authored-by: Esteban Munoz Facusse <esteban.230@hotmail.com>