fix: New overflow items all always dispatch updates to subscriber (#26565)
* fix: New overflow items might not dispatch updates

The dispatch process is done by analyzing the tops of the visibile and
invisible queues. The update triggered after adding a new element might
not be dispatched to the subscriber because the queue tops don't
change. This can result in a desync between the visibility and the
subscriber state.

Adds a `forceDispatch` flag which is used when new items are added while
observing so that the next batched update will flush to subscribers even
if queue tops don't change.

Fixes #23555

* changefile

* Add tests

* changefile

* remoive old changefile

* refactor processOverflowItems

* make test cleaner