revert: rollback List rendering change (#26896)
* Revert "Update List to render children on first render() call (#25331)"

This rendering change caused layout issues in TilesList and a performance
regression for a partner.

This reverts commit 2a970a73a77e5064e82bc9e78242e93b751e3d09.

* change files

* re-add "hasMounted" to List state

This state was added in the change being reverted. Since this interface
is exported it's part of the public API it must be maintained to avoid
a breaking change.

This state is currently not used but will be used when the feature
being reverted is re-implemented.