fix: avoid infinite loop of setStates in Sticky/ScrollablePane (#29054)
* fix: avoid infinite loop of setStates in Sticky/ScrollablePane

Background: https://github.com/microsoft/fluentui/issues/16654

The root of this issue is that certain measurement APIs return fractional
values and these values seems to jitter for display settings where the
multiples of system display scaling and browser zoom levels don't lend
themselves to returning interger values.

In this case querying `scrollTop`
would return different fractional values within a pixel of one another but
above and below an interger (e.g., 4.99 and 5.6). Critically, applying
Math.ceil(), Math.floor() or Math.round() cannot correct for this situation
in all cases. For example 4.99 becomes 5, 4 and 5, respectively while 5.6 becomes
6, 5 and 6 respectively :(

The solution is to add a wrapper for querying `scrollTop` that will treat values
within a certain range as the same. So we can get `getScrollTop(element, 1)` and
so long as the values are within a pixel of each other we will return the same
value, stabilizing the value within a small range.

* add Math.round()

* add range comparison

Because some measurement APIs return rounded values
and some require decimal values and display scaling can
affect these things we need a more permissive comparison
function.

* Apply suggestions from code review

Co-authored-by: Makoto Morimoto <Humberto.Morimoto@microsoft.com>

---------

Co-authored-by: Makoto Morimoto <Humberto.Morimoto@microsoft.com>